Cleaner commit with new activity for Login
