+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Dadbod</string>
+    <string name="auth_failed" />
 </resources>
